refactor(playback): split span-clock / LoopUnit / unit-render block into GhostPlaybackLogic.SpanClock.cs
